:seedling: Add XML support to en/disable stacktrace package detail
Changed the default enable status of packaging info in stacktraces
to be disabled to avoid the expensive lookup unless requested.

Added an XML attribute "packageTrace" to <configuration> in order to
allow enabling or disabling the packaging info in stacktraces. When
the attribute is omitted, the default setting is assumed (disabled).

To enable package detail in logback.xml:

  <configuration packageTrace="true">
    ...
  </configuration>

LOGBACK-730
LOGBACK-966