:seedling: :zap: Fix TrivialConfiguratorTest
The SaxEventRecorder is implemented with XmlPullParser in Android
(which lacks javax.xml.parsers.SAXParser that original logback uses).
XmlPullParser does not throw a SAXException for malformed XML (but
still propates an error), which allows the events to pass to the
interpreter, where an error is logged to the status manager. That
extra message throws off the unit test, which expects a specific
number of status messages.

This patch removes that status-count check, since it's not really
important in this unit test. The unit test should only check for
the XML parser error in the status list. This patch modifies the
status-list check to use Hamcrest matchers.